test(RHTAPWATCH-1174): Test on Kind
Making some changes to enable running or envtest-based tests on
a local Kind cluster. This is a stepping stone to running on CRC.

For the most part the changes have to do with dealing with some
resources already being in the cluster the 2nd time we run the tests
and with keeping test data around so we can look at it.