fix(RHTAPWATCH-1174): finalizer update permissions
We had a failure scenario  that was only happenning on OpenShift where
we couldn't set the `blockOwnerDeletion` field in `ownerReferences`
lists if the controller did not have permissions to update finalizers on
the referenced owner object kind.

Here we add the needed RBAK permissions as well as added a test that
would help prevent causing the same issue in the future.

Unfurtunately we can't reproduce the issue directly in envtest as it
seems to be unique to OpenShift.