feat(RHTAPWATCH-1192): Server-Side-Apply template
Use server-side-apply for templated objects instead of create or update
operations. This causes fields on existing resources that do not have
values sepcified in the template to be preserved.

The make the controller not clobber fields such as `spec.conteinarImage`
and `metadata.finalizers` that are set by other controllers uless
values are explicitly set in the template.