chore: reconcile on project and template changes
By having the controller watch project and template changes, dev streams
will be reconciled when those resources are created or modified.

With that in place, resources can now be created and modified in any
order, and resources will be generated per the latest changes and
owner ref will be updated between dev stream and project.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>